requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
decouple==0.0.7
Django==4.1.2
execnet==1.9.0
factory-boy==3.2.1
Faker==15.0.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-mock==3.10.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
